Add FireworksStartHook to enable fireworks without high score
Added a hook on ResultsViewController::DidActivate so that if fireworks were skipped due to not getting a high score, they will be activated anyway.
This will make it much easier to mod the fireworks as I no longer have to finish a new level or beat a record every time I need to test them.