Refactor project to use fireworkManager
Remove many of the scattered hooks and create a centralized system to control fireworks. Add everything to the TooManyFireworks namespace.
This will ensure the project scales nicely without hooks conflicting with each other. It also provides a way for fireworks to be updated anywhere anytime, such as in the config menu.