Test setting undefined UnityEngine::ParticleSystem::EmissionModule::rateOverTimeMultiplier with an icall
Test how an icall can resurrect missing functions that have been excluded from bs-cordl headers.
The EmissionModule is missing nearly all of its functions seen in a decompilation, including the functions to adjust the bursts / number of sparks. Testing with rateOverTimeMultiplier since the other half of that system, rateOverTime, is still defined.