Add number of sparks setting
Brought the ParticleSystem::EmissionModule back to life with icalls and added a setting to change the burst count from 0 to 10000.
The number of sparks NEEDS to be configurable. How could it not be? 10000 might be a little high but I can guarantee everyone will want to try it anyway. Removing sparks completely actually makes some pretty cool ambient lights.