Add CreateExtraIncrementSetting for the spawn range modal
Added a function that will take a mod config float or Vector3 variable and create UI that can change it by 10, 1, 0.1, or 0.01 increments or reset it. Spawn range modal now uses two of these for center and size.
Normal slider settings have limits and are hard to adjsut precisely, and normal increment settings can only increment by a constant amount. Instead of trying to find a balance, just add a button for everything!