Add logical limits on min and max sliders
Minimum will automatically decrease if maximum is set lower, and maximum will increase if minimum is set higher. Created CreateMinMaxConfigSliderSetting(...) and used for both the frequency and size sliders.
It is a little unintuitive for minimum to be higher than maximum or vice versa. Making one slider control two also makes them faster to adjust.