Fireworks enable and disable when opening and closing the menu
Fireworks automatically enable when the mod configuration menu is opened, and automatically disable once closed. This will replace the debug start and stop buttons.
The start and stop buttons are clunky. Still useful for testing how the fireworks behave in different menus, but it's poor user experience, especially considering how they're at the bottom of the menu.