Update template & change mod name
Update CMakeLists.txt and other files to the latest versions in Laurie's template. Also removed spaces and dashes from the mod name and ID.
This fixes an issue with setting the MOD_ID define that would lead to config being stored in Config/.json instead of Config/too-many-fireworks.json. This new naming convention is more consistent with other mods and different strings within this mod.