Split fireworksController object into solo and multiplayer versions
Refactored some things to work on a per FireworksController basis.
The multiplayer fireworks use a different object than the general ones. Both must be stored.
This commit will need to be reverted. It helped discover that multiplayer results spawn a new FireworksController every time, which is easier to implement a fix for from the previous commit than this one that has now structured things around there being two permanent FireworksControllers.