Add resultsRequireFullCombo setting
Added a setting to only show fireworks in the results screen when, in addition to the other preferences, a full combo was achieved. Also refactored levelResults.cpp a little to allow for more conditions to be added in the future.
High score isn't the only thing to be excited about!