Add enableInMainMenu setting
Added an option to enable fireworks at MainMenuViewController::DidActivate. This is flawed in that fireworks will only start going from the main menu to the level selection and not from results to level selection, but adding another option for level selection should help fix this.
Fireworks constrained to the results screen? That's enough fireworks. This is TOO MANY FIREWORKS AHHHHHH look at me break the promise I made to myself to make sane commit descriptions :P