Add enableInLevelSelection menu
Added the option to enable fireworks while in the level selection menu. Also added a few other checks to help ensure the fireworks are enabled or disabled in the proper areas.
This is considered a separate option from the main menu because fireworks may be cool to have in the general menus, but they may be distracting and noisy while trying to listen to song previews and look at level details.