Fix main fireworks being carried from one scene to another
Added despawnOnSceneTransitionHotfixHook which will despawn all of the active fireworks whenever the active scene is changed.
Fireworks continuing to exist even after magically teleporting into a different plane of existence didn't look right. This also meant that if fireworks were enabled in level selection but not gameplay, they would show for a second in gameplay which is a HUGE no no. Gameplay is not to be meddled with if not wanted.