Add main branch to workflow
Added main to the list of branches that the workflow will will run on.
Originally the dev branch was used for in-progress updates so main can represent the most recent version, but tags and version control do that just fine, and I'm not aiming to collaborate. Might as well work on main.